+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.9] - 2026-02-16
+
+### Added
+
+- Added support for enum.Enum types in Serializable
+- Added support for custom to_dict() method in Serializable
+
+### Changed
+
+- Improved type handling in Serializable
+- Replaced json_encoders with field_serializer in Serializable
+- Changed ValueError to AssertionError in SVar.validate()
+- Changed value access in SVar, now the values are stored directly in the context
+
 ## [0.1.8] - 2026-02-09
 
 ### Added
@@ -137,6 +151,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Python 3.9+ support
 - Apache-2.0 license
 
+[0.1.9]: https://github.com/alkndoom/stagecraft/compare/v0.1.8...v0.1.9
 [0.1.8]: https://github.com/alkndoom/stagecraft/compare/v0.1.7...v0.1.8
 [0.1.7]: https://github.com/alkndoom/stagecraft/compare/v0.1.6...v0.1.7
 [0.1.6]: https://github.com/alkndoom/stagecraft/compare/v0.1.5...v0.1.6